Reduce number if very small code files
